refactor: moved OIDC test fixtures to @dereekb/firebase-server/test
- extracted performFullOAuthFlow, cookie jar, and test ID token helpers into reusable oidc.test.flow.ts
- added oAuthAuthorizedSuperTestContextFactory with configurable scopes (auto-resolved from OidcAccountService)
- reordered firebase-server build so oidc builds before test sub-package
