+++ packages/dbx-firebase/oidc/package.json
@@ -14,6 +14,7 @@
     "@dereekb/firebase": "13.6.17",
     "@dereekb/rxjs": "13.6.17",
     "@dereekb/util": "13.6.17",
+    "@dereekb/vitest": "13.6.17",
     "@ngx-formly/core": "git+https://git@github.com/dereekb/ngx-formly#996d1041c8d2afbe429985a5ad394e59327bfa1d",
     "rxjs": "^7.8.0"
   }
