+++ packages/dbx-firebase/package.json
@@ -15,6 +15,7 @@
     "@dereekb/model": "13.6.17",
     "@dereekb/rxjs": "13.6.17",
     "@dereekb/util": "13.6.17",
+    "@dereekb/vitest": "13.6.17",
     "@ngrx/component-store": "^21.0.0",
     "@ngx-formly/core": "git+https://git@github.com/dereekb/ngx-formly#996d1041c8d2afbe429985a5ad394e59327bfa1d",
     "firebase": "^12.0.0",
