+++ packages/dbx-form/quiz/package.json
@@ -8,6 +8,7 @@
     "@dereekb/dbx-web": "13.6.17",
     "@dereekb/rxjs": "13.6.17",
     "@dereekb/util": "13.6.17",
+    "@dereekb/vitest": "13.6.17",
     "@ngrx/component-store": "^21.0.0",
     "rxjs": "^7.8.0"
   }
