+++ packages/dbx-web/mapbox/package.json
@@ -9,6 +9,7 @@
     "@dereekb/dbx-web": "13.6.17",
     "@dereekb/rxjs": "13.6.17",
     "@dereekb/util": "13.6.17",
+    "@dereekb/vitest": "13.6.17",
     "@ngrx/component-store": "^21.0.0",
     "@placemarkio/geo-viewport": "^1.0.2",
     "mapbox-gl": "^3.10.0",
