+++ packages/dbx-web/package.json
@@ -13,6 +13,7 @@
     "@dereekb/dbx-core": "13.6.17",
     "@dereekb/rxjs": "13.6.17",
     "@dereekb/util": "13.6.17",
+    "@dereekb/vitest": "13.6.17",
     "@ngbracket/ngx-layout": "^21.0.0",
     "@ngrx/component-store": "^21.0.0",
     "@ngrx/effects": "^21.0.0",
