+++ packages/dbx-web/table/package.json
@@ -11,6 +11,7 @@
     "@dereekb/dbx-web": "13.6.17",
     "@dereekb/rxjs": "13.6.17",
     "@dereekb/util": "13.6.17",
+    "@dereekb/vitest": "13.6.17",
     "@ngrx/component-store": "^21.0.0",
     "date-fns": "^4.0.0",
     "ngx-infinite-scroll": "^21.0.0",
