+++ packages/firebase-server/package.json
@@ -64,6 +64,7 @@
     "nanoid": "^5.1.7",
     "oidc-provider": "^9.7.0",
     "rxjs": "^7.8.0",
+    "supertest": "^7.2.2",
     "ts-essentials": "^10.0.0"
   }
 }
