+++ packages/firebase-server/test/package.json
@@ -6,6 +6,7 @@
     "@dereekb/date": "13.6.17",
     "@dereekb/firebase": "13.6.17",
     "@dereekb/firebase-server": "13.6.17",
+    "@dereekb/firebase-server/oidc": "13.6.17",
     "@dereekb/model": "13.6.17",
     "@dereekb/nestjs": "13.6.17",
     "@dereekb/rxjs": "13.6.17",
@@ -18,7 +19,8 @@
     "firebase-functions": "^7.0.0",
     "firebase-functions-test": "3.4.1",
     "jsonwebtoken": "^9.0.0",
-    "make-error": "^1.3.0"
+    "make-error": "^1.3.0",
+    "supertest": "^7.2.2"
   },
   "devDependencies": {
     "@dereekb/nestjs": "13.6.17"
