+++ packages/firebase-server/test/src/lib/index.ts
@@ -1,3 +1,4 @@
 export * from './firebase';
 export * from './firestore';
+export * from './oidc';
 export * from './storage';
