+++ packages/nestjs/eslint/package.json
@@ -3,5 +3,12 @@
   "version": "13.6.17",
   "peerDependencies": {
     "@typescript-eslint/utils": ">=8.0.0"
+  },
+  "devDependencies": {
+    "@angular/core": "21.2.3",
+    "@nestjs/common": "^11.1.17",
+    "@nestjs/core": "^11.1.17",
+    "@typescript-eslint/parser": "8.56.1",
+    "eslint": "9.39.3"
   }
 }
